update namespaces and folder structure
