+++ config/packages/security.yaml
@@ -15,7 +15,7 @@ security:
             pattern: ^.*/api/(?:latest|beta|v[0-9]+|v[0-9]+\.[0-9]+)/platform/versions$
             security: false
         authentication:
-            pattern: ^.*/authentication/providers/configurations(.*)$
+            pattern: ^.*(?<!administration)/authentication/providers/configurations(.*)$
             security: false
         api:
             stateless: true
