--- config/routes/Centreon/administration/security_policy.yaml
@@ -1,11 +1,0 @@
-centreon_security_security_policy_find_security_policy:
-    methods: GET
-    path: /administration/authentication/local/security-policy
-    controller: 'Core\Infrastructure\Security\Api\FindSecurityPolicy\FindSecurityPolicyController'
-    condition: "request.attributes.get('version') >= 22.04"
-
-centreon_security_security_policy_put_security_policy:
-    methods: PUT
-    path: /administration/authentication/local/security-policy
-    controller: 'Core\Infrastructure\Security\Api\UpdateSecurityPolicy\UpdateSecurityPolicyController'
-    condition: "request.attributes.get('version') >= 22.04"
